Only wait up to 500ms in llist destroy as there seem to exist deadlock conditions...
git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@5333 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ globals.h
@@ -1684,7 +1684,7 @@ extern void http_srv();
 
 // oscam-garbage
 #ifdef WITH_DEBUG
-extern void add_garbage_debug(void *data, char *file, int32_t line);
+extern void add_garbage_debug(void *data, char *file, uint16_t line);
 #define add_garbage(x)	add_garbage_debug(x,__FILE__, __LINE__)
 #else
 extern void add_garbage(void *data);
